Give health on Kill Enemy
Ensure that health is restored when an enemy is killed.
Also, toggle the "give health shot" option when updating the HUD to keep the HUD updated.
This would allow the HUD to be updated in a way that does not detract from the player's experience.

HUD Updated tech by K4ryuu
https://discord.com/channels/1160907911501991946/1160907912445710482/1173076426753331283